% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_data.R
\docType{data}
\name{student_data}
\alias{student_data}
\title{Generate student_data sample data}
\format{An object of class \code{data.frame} with 9000 rows and 10 columns.}
\usage{
data(student_data)
}
\description{
Generate student_data sample data
}
\examples{

data(student_data)
}
\keyword{datasets}
